# server/model/system/response/sys_authority.go
package response

import "cooller/server/model/system"

type SysAuthorityResponse struct {
	Authority system.SysAuthority `json:"authority"`
}

type SysAuthorityCopyResponse struct {
	Authority      system.SysAuthority `json:"authority"`
	OldAuthorityId int                 `json:"oldAuthorityId"` // 旧角色ID
}